Cleanup: use getDiveSelection() for shifting time
There are two cases where dive-times are shifted: in an explicit
dialog and when editing the date/time of a dive.

In each of these cases, the selected dives were collected manually.
Instead use the getDiveSelection() function. Since this returns
an std::vector, change the argument of Command::ShiftTime() to
std::vector.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>