mobile list model: add special case handling for trip expansion changing indices
This is fun. If we use swiping the dive details to change dives, there's an odd
special case. We switch from a top level dive (so all trips are collapsed) to
the last dive in the trip above it. Now that trips expands and the previously
shown dive changes its index in the dive list.

The logic to account for that looks bogus because there are so many other cases
that it doesn't deal with. But in all of them the previous index becomes invalid
as its trip is collapsed, so we don't care.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>