First version of EditDive command
Command that just swaps two dives. This restores the old QMLManager
code, which might be seen as an advantage, since it means that our
diff to master is smaller.

It is, however, a step back, as now changes to the dive sites
are not included in the undo machinery anymore. Will look at
that next.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>