mobile/dive-list: expand the trip that a dive got added to
When we are adding below, we walk forward until we hit the next trip (which is
by definition the one that the dive was added to) and expand it. Then we
continue walking until we find our dive again and make sure that it is the
current dive as far as the model/view is concerned.

For adding the dive above it should be the same logic in the other direction,
but it's actually much trickier. First, since our dive moved into a trip, the
currentIndex now already point at the dive or trip below that dive, so we need
to start the search one element higher. And then, when we find the first trip
and expand it, we need to turn around and search forward or down again.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>

make divelist search bidirectional

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>