MobileSwipeModel: improve move
That was just completely broken. Moving still confuses the model, but
this is a step in the right direction.

As for the MobileListModel, let's introduce the concept of "range"
and convert full ranges. This is needed for example when removing
an empty trip, which turns into a no-op. And, as we know by now,
Qt model/view system doesn't handle empty ranges gracefully(!).

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>